+++ strider/rectangle.py
@@ -33,24 +33,24 @@ class Rectangle:
                or self.offset_x + self.size_x > master.offset_x + master.size_x \
                or self.offset_y + self.size_y > master.offset_y + master.size_y
 
-    def move(self, x_off, y_off, master: 'Rectangle' = None):
+    def move(self, x_off, y_off, master: 'Rectangle' = None)->'Rectangle':
         ret = type(self)(self.offset_x + x_off, self.offset_y + y_off, self.size_x, self.size_y)
         if master and ret.breaks_bounds(master):
             return self
         return ret
 
-    def __floordiv__(self, other):
+    def zoom_to_center(self, other: float):
         """
         Returns a new rectangle, one nth the size on each side and centered on the original rectangle
         """
         offset_coff = (other - 1) / (2 * other)
         ret = type(self)(self.offset_x + self.size_x * offset_coff, self.offset_y + self.size_y * offset_coff,
-                          self.size_x // other, self.size_y // other)
+                         self.size_x // other, self.size_y // other)
         if min(ret.size) < 1:  # prevent a rectangle with 0 size
             return self
         return ret
 
-    def __truediv__(self, other: 'Rectangle'):
+    def size_ratio(self, other: 'Rectangle'):
         """
         calculate the ratio of sizes between self and other.
         already assumes the two rectangles have the same ratio, and only uses x
@@ -58,7 +58,7 @@ class Rectangle:
         """
         return self.size_x / other.size_x
 
-    def undiv(self, other, master: 'Rectangle' = None):
+    def zoom_from_center(self, other, master: 'Rectangle' = None)-> 'Rectangle':
         """
         note, if the new rectangle is outside master *in any way*, it returns master instead
         the only guarantee is that if R < M: